Initial starter code to Branch
Added the function and respective routes for the parts that I'm responsible for. For now, they are only placeholders. Also added a testing HTML file called 'car_id.html' to test my HTML code for the design of the page. Also made a CSS file to style that page. The HTML file has basic template and the CSS file is empty for now until next commit.