Did some Image Testing
Tried to crop the image based on it's original size so it wouldn't look as if it was squeezing to fit into the div. So I added 2 more images of varying length and width to see how it would look. Unfortunately, I didn't get to solve that. Will come back to it later.