Added well needed Comments
Comments are all over the place. Can't go wrong with anything as long we follow the comments guidelines. Everybody has a section to add their code as well, so there shouldn't be any merging issues. No change was done to the actual functions or routes. Just comments here and there.